requirements.txt
psycopg2-binary>=2.9.0
requests>=2.28.0
python-dotenv>=0.19.0
